chore: extract proxy model
